crypto: add method for applying the taptweak
The wallet returns an untweaked internal key for taproot outputs. If the
output commits to a tree of scripts, this key needs to be tweaked with
the merkle root. Even if the output does not commit to a tree of
scripts, BIP341/342 recommend commiting to a hash of the public key.

Previously, this logic for applying the taptweak was implemented in the
::SignSchnorr method.

Move this logic to its own method, ::ApplyTapTweak and refactor
::SignSchnorr to use this new method.

The motivation is to be able to use the ::ApplyTapTweak logic outside of
signing, e.g. in silent payments when retreiving the private key before
ECDH.

Note: while this is mostly a move-only commit, it is not strictly a
refactor. If signing with the merkle_root passed as an argument,
::SignSchnorr will now require two calls to `secp256k1_keypair_create`.
While this is not a big difference, its also not negligible.