wallet: check for self-transfer when sending
Call IsMineSilentPayment when sending to a SP address if our wallet has
SPs enabled.

In `CreateSilentPaymentOutputs`, we create a change output using silent payments. The
presence of the change output will cause us to not fully check the
transaction and thus never create the silent payment tweaks, which is
why we check for self-transfers here. This feels a bit hacky, but works
for now.