add TweakAdd method to CPubKey
this is a temporary measure. The alternative to having this method is
to store the full spk + tweak in the DB (34 + 32 bytes), vs only storing
the tweak (32 bytes) and recreating the spks every time we load the tweaks.

Another option is to add this method directly to V0SilentPayments so that
its not generally available for any CPubKey.