wallet: use CRecipient instead of CTxOut
Now that a CRecipient holds a CTxDestination, we can get the serialized
size and determine if the output is dust using a visitor. This does not
change any current behavior, but provides a nice generalization that can
be used to apply special logic to a CTxDestination serialization and dust
calculations in the future.