wallet: use CRecipient instead of CTxOut
Now that a CRecipient holds a CTxDestination, we can get the serialized
size and determine if the output is dust using a visitor. This does not
change any current behavior, but provides a nice generalization that can
be used to apply special logic to a CTxDestination serialization and dust
calculations in the future.

Specifically, the visitor is needed here because in a later PR when
support for `V0SilentPayment` destinations is added, we need to use
`WitnessV1Taproot` as the scriptPubKey for serialized size calcuations
whenever the `CRecipient` destination is a `V0SilentPayment` destination
(recall that `CTxDestination` is a std::variant).