kernel: Add validation interface and task runner to C header
This adds the infrastructure required to process validation events. For
now the external validation interface only has support for the
`BlockChecked` callback, but support for the other internal validation
interface methods can be added in the future.

Both the validation interface and the task runner follow a architecture
for defining their callbacks and ownership that is similar to the
notifications.

The task runner is registered with a context, which internally creates a
unique ValidationSignals object. When the user creates a new chainstate
manager the validation signals are internally passed to the chainstate
manager through the context.

The user has great flexibility when defining the callbacks of the task
runner, allowing for immediate, asynchronous, or threaded processing of
the inserted validation events. A validation event has to be processed
through the `kernel_execute_event_and_destroy` function.

A validation interface can register for validation events with a
context. Internally the passed in validation interface is registerd with
the validation signals of a context.

The BlockChecked callback introduces a seperate type for a non-owned
block. Since a library-internal object owns this data, the user needs to
be explicitly prevented from deleting it. In a later commit a utility
will be added to copy its data.