remove labels
this isnt really how we want labels implemented, i.e., getnewaddress should only
return a new address if the user explicitly requests a label.

labels overall are a bit trickier to implement in the descriptor wallet. its unclear
to me if it should happen at the CWallet level, the DesSPK level, or even the Desc level.