refactor: use KeyPair in SignSchnorr
Use `KeyPair` instead of creating a `secp256k1_keypair` object. The
main change here is creating a `KeyPair` instead of a
`secp256k1_keypair` and then passing it to the libsec256k1 functions
using `reinterpret_cast<secp256k1_keypair*>(keypair)`.

The variable name `keypair` is used for the reinterpret_cast to simplify the
diff in a later commit when all of the logic in SignSchnorr is moved into the
KeyPair class.

Note: we no longer need to call memory_cleanse since `KeyPair` is now
using a secure allocator (same as CKey). See src/support/allocator/secure.h