More sp tests (#5)
* test: add wallet persistence

currently failing for closing / reopening wallets with a payment.
does not fail closing / reopening an empty silent payments wallet.

* wallet: fix error loading tweaks from db

Error was caused by the wallet not creating a SilentPaymentDescriptorScriptPubKeyMan for sp desc

* wallet: fix amount for sp payments not loaded corrrectly

The wallet tries to topup the keypool for the descspkmans after loading.
This change overrides the DescriptorScriptpubPubKeyMan::TopUp to add spks from received silent payments to wallet spks cache (m_cached_spks) so amount and tx details can be loaded correctly

* test: add silent payment to wallet descriptor tests

- Always set `internal` to `false` for sp descriptors in `listdescriptors` list
- On `getaddressinfo` for silent payment address, search for scan key among wallet silent payment spkmans
- Return the same silent payment for internal and external use

* test: wallet import rescan for silent payments

* test: Silent Payments RBF

* rpc/generatetoaddress: Reject Silent Payment addresses

* Set string for sp outputtype to "silent-payments"

* fix build error caused by mutex

* Remove unused function

* Fix lint error

* Fix lint error

* Enable silent_payments in fuzz test wallet

---------

Co-authored-by: josibake <josibake@protonmail.com>