test: call IsMine inside AddTx
Call `IsMine(tx, coins)` inside the AddTx helper function to
update the scriptpubkey cache with self-transfer silent payment outputs.

This is not needed for traditional scripts since they are already known
to the wallet, so self transfers are easily detectable. In the case of
silent payments, the scripts are not known until the transaction is created,
so calling IsMine on the new transaction ensures we can detect a self transfer
without needing to go through the block connected or in mempool notification events.