descriptors: disable addr(<silent-payment-address>)
A Silent payment address consist of a public scan key and public spend key.

This is enough information to construct a tx paying to the address, but not enough information to find payments made to that address.

Since descriptors are meant to contain information needed for a wallet to find payments and/or spend them, having an addr(<silent-payment-address>) descriptor is useless